% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{accu_plot}
\alias{accu_plot}
\title{Plot accumulation curves for \code{\link{phyloseq-class}} object}
\usage{
accu_plot(physeq, fact = NULL, nbSeq = TRUE, step = NULL,
  by.fact = FALSE, ci.col = NULL, col = NULL, lwd = 3, leg = TRUE,
  printSamplesNames = FALSE, CI = 2, ...)
}
\arguments{
\item{physeq}{(Required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(Required): Name of the factor in physeq@sam_data used to plot different lines}

\item{nbSeq}{(logical): Either plot accumulation curves using sequences or using samples}

\item{step}{(integer): distance among points calculated to plot lines. A
low value give better plot but is more time consuming. Only use if nbSeq = TRUE}

\item{by.fact}{(logical): First merge the OTU table by factor to plot only one line by factor}

\item{ci.col}{: Color vector for confidence intervall. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.}

\item{col}{: Color vector for lines. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.}

\item{lwd}{(default = 3): thickness for lines. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.}

\item{leg}{(logical): Plot legend or not. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.}

\item{printSamplesNames}{(logical): Print samples names or not? Only use if nbSeq = TRUE.}

\item{CI}{(default = 2) : Confidence intervall value used to multiply the standard error to plot confidence intervall}

\item{...}{Additional arguments passed on to \code{\link{ggplot}} if nbSeq = TRUE 
or \code{\link{plot}} if nbSeq = FALSE}
}
\value{
A \code{\link{ggplot}}2 plot representing the richness 
accumulation plot if nbSeq = TRUE, else, if nbSeq = FALSE
return a base plot.
}
\description{
Plot accumulation curves for \code{\link{phyloseq-class}} object
}
\examples{
data("GlobalPatterns")
GlobalPatterns_Archaea <- subset_taxa(GlobalPatterns,
 GlobalPatterns@tax_table[, 1] == 'Archaea')
\dontrun{
accu_plot(GlobalPatterns_Archaea, 'SampleType', nbSeq = TRUE, by.fact = TRUE)
}

}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[vegan]{specaccum}}
}

