% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_filter-methods.R
\name{as.binaryOtuTable}
\alias{as.binaryOtuTable}
\title{Transform the otu_table of a \code{\link{phyloseq-class}} object into a binary otu_table. Usefull to test if the results are not biaised by sequences bias that appended during PCR or NGS pipeline.}
\usage{
as.binaryOtuTable(physeq, minNumber = 1)
}
\arguments{
\item{physeq}{(Required): a \code{\link{phyloseq-class}} object.}

\item{minNumber}{(default = 1): the minimum number of sequences to put 
a 1 in the otu table.}
}
\value{
A \code{\phyoseq} object with only 0/1 in the OTU table
}
\description{
Transform the otu_table of a \code{\link{phyloseq-class}} object into a binary otu_table. Usefull to test if the results are not biaised by sequences bias that appended during PCR or NGS pipeline.
}
\examples{
data(enterotype)
#enterotype.bin <- as.binaryOtuTable(enterotype)
}
\author{
Adrien Taudiere
}

