\docType{methods}
\name{otuTable<-}
\alias{assign-otuTable}
\alias{otuTable<-}
\alias{otuTable<-,phyloseq,otuTable-method}
\title{Assign to otuTable an object/value.}
\usage{
  otuTable(x) <- value
}
\arguments{
  \item{x}{(Required). The object within which you will
  replace something}

  \item{value}{(Required). The value with which you will
  replace that thing in x.}
}
\description{
  Assign to otuTable an object/value.
}
\examples{
#
}

