\name{sampleMap<-}
\alias{assign-sampleMap}
\alias{sampleMap<-}
\title{Assign to sampleMap an object/value.}
\usage{
  sampleMap(x) <- value
}
\arguments{
  \item{x}{(Required). The object within which you will
  replace something}

  \item{value}{(Required). The value with which you will
  replace that thing in x.}
}
\description{
  Assign to sampleMap an object/value.
}
\examples{
#
}

