\name{taxTab<-}
\alias{assign-taxTab}
\alias{taxTab<-}
\title{Assign to taxTab an object/value.}
\usage{
  taxTab(x) <- value
}
\arguments{
  \item{x}{(Required). The object within which you will
  replace something}

  \item{value}{(Required). The value with which you will
  replace that thing in x.}
}
\description{
  Assign to taxTab an object/value.
}
\examples{
#
}

