\docType{methods}
\name{tre<-}
\alias{assign-tre}
\alias{tre<-}
\alias{tre<-,otuTree,phylo-method}
\alias{tre<-,otuTree,phylo4-method}
\title{Assign to tre an object/value.}
\description{
  This will automatically convert "phylo"-class trees (ape
  package) to "phylo4"-class trees (phylobase package).
}
\examples{
#
}

