\docType{data}
\name{data-ex1}
\alias{data-ex1}
\alias{ex1}
\title{Example data set 1 for the phyloseq package.}
\description{
  Example data set 1 for the phyloseq package.
}
\author{
  Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\references{
  \url{www.stanford.edu/~mcmurdie}
}
\keyword{data}

