% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{dist-class}
\alias{dist-class}
\title{An S4 placeholder for the \code{\link[stats]{dist}} class.}
\description{
See \code{\link[stats]{dist}} for details
about this type of a distance matrix object.
}
\seealso{
\code{\link[stats]{dist}}, \code{\link{setOldClass}}
}

