% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{dpcoa}
\alias{dpcoa}
\title{S3 class placeholder definition (list) for dpcoa}
\format{An object of class \code{dpcoa} of length 0.}
\usage{
dpcoa
}
\description{
The ade4 package does not export a version of its \code{\link[ade4]{dpcoa}}-class,
partly because it is not really defined formally anywhere.
Instead, it is an S3 class extended from the base class, \code{\link{list}} --
this is a very common and easy approach --
and proper behavior of any method taking an instance of this class 
requires exact naming conventions for element names of the list components.
The phyloseq package does not provide any validity checks that a given phylo
instance is valid (conforms to the conventions in the ape package). Yet.
If problems arise, this might be considered, and they could be defined
judiciously and within phyloseq.
}
\details{
An instance of this class can be produced from within phyloseq using either
the \code{\link{DPCoA}} function, or the higher-level wrapping function
\code{\link{ordinate}}.
}
\seealso{
\code{\link[ade4]{dpcoa}}

\code{\link{DPCoA}}

\code{\link{ordinate}}
}
\keyword{internal}
