% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo-class.R
\docType{methods}
\name{fix_phylo}
\alias{fix_phylo}
\alias{fix_phylo,phylo-method}
\title{Method for fixing problems with phylo-class trees in phyloseq}
\usage{
fix_phylo(tree)

\S4method{fix_phylo}{phylo}(tree)
}
\description{
For now this only entails replacing each missing (\code{NA}) branch-length
value with 0.0.
}
\keyword{internal}

