\name{get.component.classes}
\alias{get.component.classes}
\title{Get the component objects classes and slot names.}
\value{
  a character vector of the component objects classes,
  where each element is named by the corresponding slot
  name in the higher-order phyloseq objects (objects
  containing more than 1 phyloseq data object).
}
\description{
  Get the component objects classes and slot names.
}
\examples{
#
#get.component.classes()
}
\seealso{
  merge_phyloseq
}
\keyword{internal}

