% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{hitchip2physeq}
\alias{hitchip2physeq}
\title{hitchip2physeq}
\usage{
hitchip2physeq(otu, meta, taxonomy = NULL, detection.limit = 10^1.8)
}
\arguments{
\item{otu}{Sample x OTU absolute HITChip signal}

\item{meta}{Sample x features metadata data.frame}

\item{taxonomy}{OTU x Taxonomy data.frame (HITChip taxonomy used by default)}

\item{detection.limit}{HITChip signal detection limit (absence / presence)}
}
\value{
phyloseq object
}
\description{
Convert HITChip data into phyloseq format
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}

