% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{import_hitchip}
\alias{import_hitchip}
\title{Import HITChip output into phyloseq format}
\usage{
import_hitchip(data.dir, method = "rpa", detection.threshold = 0,
  verbose = F)
}
\arguments{
\item{data.dir}{Profiling script output directory for reading the data.}

\item{method}{Probe summarization method ("rpa" or "sum")}

\item{detection.threshold}{Taxon absence/presence thresholds (typically 10^1.8 for HITChip)}

\item{verbose}{verbose}
}
\value{
data matrix (phylo x samples)
}
\description{
Import HITChip output into phyloseq format
}
\examples{
\dontrun{
    # Define example data folder with HITChip phylogenetic microarray data
    data.dir <- system.file("extdata/hitchip", package = "phyloseq")
    # Import the microarray data in phyloseq format
    physeq <- import_hitchip(data.dir)
  }
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
To cite the import_hitchip function, use
  "Leo Lahti 2015. import_hitchip function in the phyloseq R package."
  For the rpa method for probe summarization, cite:
  Lahti et al. NAR 41(10):e110, 2013.
  URL \url{http://nar.oxfordjournals.org/content/41/10/e110}
}
\keyword{utilities}

