% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{n.phylotypes.per.oligo}
\alias{n.phylotypes.per.oligo}
\title{Description: Check number of matching phylotypes for each probe}
\usage{
n.phylotypes.per.oligo(taxonomy, level)
}
\arguments{
\item{taxonomy}{oligo - phylotype matching data.frame}

\item{level}{phylotype level

Returns:}
}
\value{
number of matching phylotypes for each probe
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}

