% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{otu_circle}
\alias{otu_circle}
\title{Plot OTU circle for \code{\link{phyloseq-class}} object}
\usage{
otu_circle(physeq = NULL, fact = NULL, taxa = "Order", nbSeq = TRUE,
  rarefy = FALSE, min.prop.tax = 0.01, min.prop.mod = 0.1,
  gap.degree = NULL, start.degree = NULL, row.col = NULL,
  grid.col = NULL, ...)
}
\arguments{
\item{physeq}{(Required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(Required): Name of the factor to cluster samples by modalities. Need to be in \code{physeq@sam_data}.}

\item{taxa}{(Default:'Order'): Name of the taxonomic rank of interest}

\item{nbSeq}{(Default: TRUE): Represent the number of sequences or the number of OTUs (nbSeq = FALSE)}

\item{rarefy}{(logical): Does each samples modalities need to be rarefy in order to compare them with the same amount of sequences?}

\item{min.prop.tax}{(Default: 0.01): The minimum proportion for taxon to be ploted}

\item{min.prop.mod}{(Default: 0.1) : The minimum proportion for modalities to be ploted}

\item{gap.degree}{: Gap between two neighbour sectors. It can be a single value or a vector. If it is a vector, the first value corresponds to the gap after the first sector.}

\item{start.degree}{: The starting degree from which the circle begins to draw. Note this degree is measured in the standard polar coordinate which means it is always reverse-clockwise.}

\item{row.col}{: Color vector for row}

\item{grid.col}{: Grid colors which correspond to sectors. The length of the vector should be either 1 or the number of sectors. It's preferred that grid.col is a named vector of which names correspond to sectors. If it is not a named vector, the order of grid.col corresponds to order of sectors.}

\item{...}{Additional arguments passed on to \code{\link[circlize]{chordDiagram}} or \code{\link[circlize]{circos.par}}}
}
\value{
A \code{\link{chordDiagram}} plot representing the distribution 
of OTUs or sequences in the different modalities of the factor fact
}
\description{
Plot OTU circle for \code{\link{phyloseq-class}} object
}
\examples{
data("GlobalPatterns")
# GlobalPatterns_Archaea <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == 'Archaea')
# otu_circle(GlobalPatterns_Archaea, 'SampleType')
}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[circlize]{chordDiagram}}

\code{\link[circlize]{circos.par}}
}

