\name{otuSamTax-class}
\alias{otuSamTax-class}
\title{Define the otuSamTax class.}
\description{
  Inherits the otuSam class. \describe{ \item{otuTable}{a
  single object of class otuTable.} \item{sampleMap}{a
  single object of class sampleMap.} \item{taxTab}{a single
  object of class taxonomyTable.} }
}

