\name{otuSamTax}
\alias{otusamtax}
\alias{otuSamtax}
\alias{otuSamTax}
\title{Subset just the otuSamTax portion of a H.O. object.}
\usage{
  otuSamTax(object)
}
\arguments{
  \item{object}{An object among the set of classes defined
  by the phyloseq package that contain a taxonomyTable,
  sampleMap, and an otuTable.}
}
\value{
  A phyloseq-package object of class \code{otuSamTax}.
}
\description{
  Subset just the otuSamTax portion of a H.O. object.
}
\examples{
#
}

