\name{otuSamTaxTree-class}
\alias{otuSamTaxTree-class}
\title{Define the otuSamTaxTree class.}
\description{
  Contains all (current) component classes: otuTable,
  sampleMap, taxonomyTable (taxTab), and phylo (tre).
  Inherits otuSamTax and otuSamTree.
}
\details{
  \describe{ \item{otuTable}{a single object of class
  otuTable.} \item{sampleMap}{ a single object of class
  sampleMap.} \item{taxTab}{ a single object of class
  taxonomyTable.} \item{tre}{ a single object of class
  phylo, from the package ape} }
}

