\name{otuSamTree-class}
\alias{otuSamTree-class}
\title{Define the otuSamTree class.}
\description{
  Contains otuTable, sampleMap, phylo (tre) slots. Inherits
  the otuSam and otuTree classes. \describe{
  \item{otuTable}{a single object of class otuTable}
  \item{sampleMap}{a single object of class sampleMap.}
  \item{tre}{a single object of class phylo, from the
  package ape} }
}

