\name{otuSamTree}
\alias{otusamtree}
\alias{otuSamtree}
\alias{otuSamTree}
\title{Subset just the otuSamTree portion of a H.O. object.}
\usage{
  otuSamTree(object)
}
\arguments{
  \item{object}{An object among the set of classes defined
  by the phyloseq package that contain a phylogenetic tree,
  sampleMap, and an otuTable.}
}
\value{
  A phyloseq-package object of class \code{otuSamTree}.
}
\description{
  Subset just the otuSamTree portion of a H.O. object.
}
\examples{
#
}

