\name{otuTree4-class}
\alias{otuTree4-class}
\title{Define the otuTree4 class.}
\description{
  Identical to otuTree class, but uses \code{phylo4} class
  for tree slot instead of \code{phylo}. \describe{
  \item{otuTable}{ a single object of class otuTable}
  \item{tre}{ a single object of class phylo4, imported
  from phylobase.} }
}

