% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{phyloS4}
\alias{phyloS4}
\title{An S4 equivalent to the \code{\link[ape]{phylo}} class in ape.}
\description{
See the \code{\link[ape]{ape}} package for details about this type of
representation of a phylogenetic tree. It is used throught ape.
}
\details{
This is an S4 placeholder of the main phylogenetic tree class from the ape package.
The ape package does not export a version of its \code{\link[ape]{phylo}}-class,
in large part because it is not defined formally anywhere.
Instead, it is an S3 class extended from a base class, \code{\link{list}}.
This is a very common and easy approach utilized in many S3 methods.
However, proper behavior of any method taking an instance of
the \code{\link[ape]{phylo}} class
requires exact naming conventions for element names of the components.
The phyloseq package now includes some validity checks on required components,
and this may improve over time as this S4/S3 transition is tested.
If a formal definition for the the phylo-class is ever exported
by ape, the current philosophy of phyloseq would be to
import those from ape rather than maintain duplicate
internal definitions. Note that there is still some
work going on for the phylobase package, which is addressing these same
exact issues for S4 phylogenetic tree interaction.
A very large number of packages (around 60 at my last count), depend on ape,
making it easily the de facto standard for representing phylogenetic trees in R;
and the phyloseq team would prefer to use any exported definitions from
the ape package if possible and available.
}
\section{Slots}{

\describe{
\item{\code{edge}}{A two-column matrix of mode numeric where each row represents
an edge of the tree; the nodes and the tips are symbolized with numbers;
the tips are numbered from 1 to the number of tips,
and the nodes are numbered after the tips.
For each row, the first column gives the ancestor.}

\item{\code{edge.length}}{(Optional) a numeric vector giving the lengths
of the branches given by \code{edge}.}

\item{\code{tip.label}}{A vector of mode character giving the names of the tips;
the order of the names in this vector corresponds to the
(positive) number in edge.}

\item{\code{Nnode}}{The number of (internal) nodes.}

\item{\code{node.label}}{(Optional). A vector of mode character
giving the names of the nodes.}

\item{\code{root.edge}}{(Optional). A numeric value giving the length
 of the branch at the root if it exists.}
}}
\seealso{
\code{\link[ape]{phylo}}, \code{\link{setOldClass}}
}

