% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_metagenomeSeq.R
\name{phyloseq_to_metagenomeSeq}
\alias{phyloseq_to_metagenomeSeq}
\title{Convert phyloseq data to MetagenomeSeq MRexperiment object}
\usage{
phyloseq_to_metagenomeSeq(physeq, ...)
}
\arguments{
\item{physeq}{(Required). \code{\link{phyloseq-class}}.}

\item{...}{(Optional). Additional named arguments passed 
to \code{\link[metagenomeSeq]{newMRexperiment}}.
Most users will not need to pass any additional arguments here.}
}
\value{
A \code{\link[metagenomeSeq]{MRexperiment-class}} object.
}
\description{
No testing is performed by this function. The phyloseq data is converted
to the relevant \code{\link[metagenomeSeq]{MRexperiment-class}} object, which can then be
tested in the zero-inflated mixture model framework
(e.g. \code{\link[metagenomeSeq]{fitZig}}) 
in the metagenomeSeq package.
See the
\href{http://joey711.github.io/phyloseq-extensions}{phyloseq-extensions}
tutorials for more details.
}
\examples{
 # Check out the vignette metagenomeSeq for more details.
 # vignette("metagenomeSeq")
 data(soilrep)
 phyloseq_to_metagenomeSeq(soilrep)
}
\seealso{
\code{\link[metagenomeSeq]{fitTimeSeries}}
 \code{\link[metagenomeSeq]{fitLogNormal}}
 \code{\link[metagenomeSeq]{fitZig}}
 \code{\link[metagenomeSeq]{MRtable}}
 \code{\link[metagenomeSeq]{MRfulltable}}
}

