\name{phyloseqFather-class}
\alias{phyloseqFather-class}
\title{Keystone S4 virtual class inherited by all other complex phyloseq classes.}
\description{
  All other complex classes inherit by virtue of the fact
  that they contain an otuTable as one of their data
  components.
}
\details{
  Slots: \describe{ \item{otuTable}{ Contains an otu
  abundance table object. See otuTable class.} }
}

