\name{phyloseqTax-class}
\alias{phyloseqTax-class}
\title{Define the phyloseqTax class.}
\description{
  Inherits the phyloseq class. \describe{ \item{otuTable}{a
  single object of class otuTable.} \item{sampleMap}{a
  single object of class sampleMap.} \item{taxTab}{a single
  object of class taxonomyTable.} }
}

