\name{phyloseqTax}
\alias{phyloseqtax}
\alias{phyloseqTax}
\title{Subset just the "phyloseqTax" portion of a H.O. object.}
\arguments{
  \item{object}{An object among the set of classes defined
  by the phyloseq package that contain a taxonomyTable,
  sampleMap, and an otuTable.}
}
\value{
  A phyloseq-package object of class \code{phyloseqTax}.
}
\description{
  Subset just the "phyloseqTax" portion of a H.O. object.
}
\examples{
#
}
\seealso{
  phyloseq
}
\keyword{subset}

