\name{phyloseqTree-class}
\alias{phyloseqTree-class}
\title{Define the phyloseqTree class.}
\description{
  Contains otuTable, sampleMap, phylo (tre) slots. Inherits
  the phyloseq and otuTree classes. \describe{
  \item{otuTable}{a single object of class otuTable}
  \item{sampleMap}{a single object of class sampleMap.}
  \item{tre}{a single object of class phylo, from the
  package ape} }
}

