\name{phyloseqTree}
\alias{phyloseqtree}
\alias{phyloseqTree}
\title{Define the phyloseqTree class.}
\arguments{
  \item{object}{An object among the set of classes defined
  by the phyloseq package that contain a phylogenetic tree,
  sampleMap, and an otuTable.}
}
\value{
  A phyloseq-package object of class \code{phyloseqTree}.
}
\description{
  Contains otuTable, sampleMap, phylo (tre) slots. Inherits
  the phyloseq and otuTree classes.

  Subset just the "phyloseqTree" portion of a H.O. object.
}
\examples{
#
}
\seealso{
  phyloseq
}
\keyword{subset}

