\docType{methods}
\name{plot}
\alias{"plot,otuTree-method"}
\alias{"plot,phyloseq-method"}
\alias{"plot,phyloseqTree-method"}
\alias{plot}
\title{Plot an otuTree object as a simple phylogenetic tree.}
\description{
  Plot an otuTree object as a simple phylogenetic tree.

  Plot a phyloseq object as a taxaplot style barplot.

  Plot a phyloseqTree object as an annotated phylogenetic
}

