% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_DESeq2.R
\name{plot_deseq2_phyloseq}
\alias{plot_deseq2_phyloseq}
\title{Plot DESeq2 results for a phyloseq or a DESeq2 object.}
\usage{
plot_deseq2_phyloseq(data, contrast = NULL, tax_table = NULL,
  alpha = 0.01, taxa = "Genus", color_tax = "Phylum", taxDepth = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{data}{(Required): a \code{\link{phyloseq-class}} or a \code{\link{DESeqDataSet-class}} object.}

\item{contrast}{(Required):This argument specifies what comparison to extract from the object to build a results table. See \code{\link[DESeq2]{results}} man page for more details.}

\item{tax_table}{: Required if data is a \code{\link{DESeqDataSet-class}} object. 
The taxonomic table used to find the \code{taxa} and \code{color_taxa} arguments. If data is a \code{\link{phyloseq-class}} object, data@tax_table is used.}

\item{alpha}{(default = 0.01): the significance cutoff used for optimizing the independent filtering. If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.}

\item{taxa}{(default = 'Genus'): taxonomic level of interest}

\item{color_tax}{(default = 'Phylum'): taxonomic level used for color or a color vector.}

\item{taxDepth}{(default = NULL): Taxonomic depth to test for differential distribution among contrast. If Null the analysis is done at the OTU (i.e. Species) level. If not Null data need to be a \code{\link{phyloseq-class}} object.}

\item{verbose}{: whether the function print some information during the computation}

\item{...}{Additional arguments passed on to \code{\link[DESeq2]{DESeq}} or \code{\link[ggplot2]{ggplot}}}
}
\value{
A \code{\link{ggplot}}2 plot representing DESeq2 results
}
\description{
Plot DESeq2 results for a phyloseq or a DESeq2 object.
}
\examples{
\dontrun{
data("GlobalPatterns")
GlobalPatterns_Proteobacteria <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 2] == 'Proteobacteria')
res_deseq2 <- DESeq(phyloseq_to_deseq2(GlobalPatterns_Proteobacteria, ~ SampleType), test = 'Wald', fitType = 'local')
plot_deseq2_phyloseq(res_deseq2, c('SampleType', 'Soil', 'Feces'), tax_table = GlobalPatterns_Proteobacteria@tax_table, color_tax = 'Kingdom')
plot_deseq2_phyloseq(res_deseq2, c('SampleType', 'Soil', 'Feces'), tax_table = GlobalPatterns_Proteobacteria@tax_table, color_tax = c("red", "black"), alpha = 0.7))
plot_deseq2_phyloseq(GlobalPatterns_Proteobacteria, c('SampleType', 'Soil', 'Feces'), taxDepth = 'Family', taxa = 'Family',  color_tax = 'Class')
}
}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[DESeq2]{DESeq}}

\code{\link[DESeq2]{results}}

\code{\link{plot_edgeR_phyloseq}}
}

