% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_edgeR.R
\name{plot_edgeR_phyloseq}
\alias{plot_edgeR_phyloseq}
\title{Plot edgeR results for a phyloseq or a edgeR object.}
\usage{
plot_edgeR_phyloseq(data, contrast = NULL, alpha = 0.01, taxa = "Genus",
  color_tax = "Phylum", verbose = TRUE, ...)
}
\arguments{
\item{data}{(Required): a \code{\link{phyloseq-class}} or a \code{\link{DESeqDataSet-class}} object.
#' @param tax_table: Require if data is a \code{\link{DESeqDataSet-class}} object. 
The taxonomic table used to find the \code{taxa} and \code{color_taxa} arguments. If data is a \code{\link{phyloseq-class}} object, data@tax_table is used.}

\item{contrast}{(Required):This argument specifies what comparison to extract from the object to build a results table. See \code{\link[DESeq2]{results}} man page for more details.}

\item{alpha}{(default = 0.01): the significance cutoff used for optimizing the independent filtering. If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.}

\item{taxa}{(default = 'Genus'): taxonomic level of interest}

\item{color_tax}{(default = 'Phylum'): taxonomic level used for color assignation}

\item{verbose}{: whether the function print some information during the computation}

\item{...}{Additional arguments passed on to \code{\link[edgeR]{exactTest}} or \code{\link[ggplot2]{ggplot}}}
}
\value{
A \code{\link{ggplot}}2 plot representing edgeR results
}
\description{
Plot edgeR results for a phyloseq or a edgeR object.
}
\examples{
\dontrun{
data(GlobalPatterns)
plot_edgeR_phyloseq(GlobalPatterns, c('SampleType', 'Soil', 'Feces'), tax_table = GlobalPatterns@tax_table, color_tax = 'Kingdom')
plot_edgeR_phyloseq(GlobalPatterns, c('SampleType', 'Soil', 'Feces'), taxa = 'Class', tax_table = GlobalPatterns@tax_table, color_tax = 'Kingdom')
}
}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[edgeR]{exactTest}}

\code{\link{plot_deseq2_phyloseq}}
}

