% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multtest-wrapper.R
\name{plot_mt}
\alias{plot_mt}
\title{Plot the result of a mt test (\code{\link[phyloseq]{mt}})}
\usage{
plot_mt(mt = NULL, alpha = 0.05, color_tax = "Class", taxa = "Species")
}
\arguments{
\item{mt}{(Required): result of a mt test}

\item{alpha}{(default = 0.05): Choose the cut off p-value to plot taxa}

\item{color_tax}{: A taxonomic level to color the points}

\item{taxa}{: The taxonomic level choose for x-positioning}
}
\value{
a \code{\link{ggplot}}2 plot of result of a mt test
}
\description{
Plot the result of a mt test (\code{\link[phyloseq]{mt}})
}
\examples{
\dontrun{
data("GlobalPatterns")
res = mt(GlobalPatterns, "SampleType", test="f")
plot_mt(res, color_tax = "Phylum") + scale_color_hue()
}
}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[phyloseq]{mt}}
}

