% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranknames2taxanames.R
\name{ranknames2taxanames}
\alias{ranknames2taxanames}
\title{convert ranknames to taxanames}
\usage{
ranknames2taxanames(taxa, x, rankby = "Genus")
}
\arguments{
\item{taxa}{A character vector of the taxa in object x that you 
want to keep. See \code{\link{prune_taxa}} for details.}

\item{x}{A phylogenetic object}

\item{rankby}{one of rank_names(x)}
}
\value{
a character vector of taxanames
data("GlobalPatterns")
rankby = "Genus"
TopNrank <- names(sort(taxa_sums(GlobalPatterns, rankby), TRUE)[1:5])
TopNOTUs <- ranknames2taxanames(TopNrank, GlobalPatterns, rankby)
rankby_prune  <- prune_taxa(TopNOTUs, GlobalPatterns)
}
\description{
convert ranknames, like Genus, to taxanames, like otu1.
}
