% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{retrieve.probesets}
\alias{retrieve.probesets}
\title{retrieve.probesets}
\usage{
retrieve.probesets(tax.table, level = "species", name = NULL)
}
\arguments{
\item{tax.table}{data.frame with oligo - phylotype
mapping information}

\item{level}{phylotype level for probesets}

\item{name}{specify phylotypes to check (optional)}
}
\value{
A list. Probes for each phylotype.
}
\description{
List probes for each probeset
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}

