\name{sample_data-class}
\alias{sample_data-class}
\title{The S4 for storing sample variables.}
\description{
  Row indices represent samples, while column indices
  represent experimental categories, variables (and so
  forth) that describe the samples.
}
\details{
  \describe{

  \item{.Data}{data-frame data, inherited from the
  data.frame class.}

  \item{row.names}{ Also inherited from the data.frame
  class; it should contain the sample names. }

  \item{names}{Inherited from the data.frame class.}

  }
}

