% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{sankey_phyloseq}
\alias{sankey_phyloseq}
\title{Sankey plot of \code{\link{phyloseq-class}} object}
\usage{
sankey_phyloseq(physeq = NULL, fact = NULL, taxa = c(1:4),
  nbSeq = FALSE, min.prop.tax = 0, tax2remove = NULL, units = NULL,
  Symbol2sub = c("\\\\.", "-"), ...)
}
\arguments{
\item{physeq}{(Required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(Optional): Name of the factor to cluster samples by modalities. 
Need to be in \code{physeq@sam_data}.}

\item{taxa}{(Default: c(1:4)): a vector of taxonomic rank to plot}

\item{nbSeq}{(Default: FALSE): Represent the number of sequences or the number of OTUs (nbSeq = FALSE). Note that ploting the number of sequences is slower.}

\item{min.prop.tax}{(Default: 0): The minimum proportion for taxon to be ploted. EXPERIMENTAL. For the moment each links below the min.prop. tax is discard from the sankey network resulting in sometimes weird plot.}

\item{tax2remove}{: a vector of taxonomic groups to remove from the analysis (e.g. \code{c('Incertae sedis', 'unidentified')})}

\item{units}{: character string describing physical units (if any) for Value}

\item{Symbol2sub}{(default = c('\\.', '-')): vector of symbol to delete in the taxonomy}

\item{...}{Additional arguments passed on to \code{\link[networkD3]{sankeyNetwork}}}
}
\value{
A \code{\link{sankeyNetwork}} plot representing the taxonomic distribution
of OTUs or sequences. If \code{fact} is set, represent the distribution of
the last taxonomic level in the modalities of \code{fact}
}
\description{
Sankey plot of \code{\link{phyloseq-class}} object
}
\examples{
data("GlobalPatterns")
#GlobalPatterns_Archaea <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == 'Archaea')
#sankey_phyloseq(GlobalPatterns_Archaea)
#sankey_phyloseq(GlobalPatterns_Archaea, fact = 'SampleType')
\dontrun{
sankey_phyloseq(GlobalPatterns, taxa = c(1:5), min.prop.tax = 0.01)
sankey_phyloseq(GlobalPatterns, taxa = c(2:6), min.prop.tax = 0.01, nbSeq = TRUE)
}
}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[networkD3]{sankeyNetwork}}
}

