\docType{methods}
\name{show}
\alias{show}
\alias{show,otuTable-method}
\alias{show,phyloseq-method}
\alias{show,sampleData-method}
\alias{show,taxonomyTable-method}
\title{method extensions to show for phyloseq classes.}
\description{
  See the general documentation of
  \code{\link[methods]{show}} method for expected behavior.
}
\examples{
# data(ex1)
# print(ex1)
# ex1
}
\seealso{
  \code{\link[methods]{show}}
}

