% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{summarize.rpa}
\alias{summarize.rpa}
\title{Description: Probeset summarization with RPA}
\usage{
summarize.rpa(taxonomy, level, probedata, verbose = TRUE,
  probe.parameters = NULL)
}
\arguments{
\item{taxonomy}{oligo - phylotype matching data.frame}

\item{level}{taxonomic level for the summarization.}

\item{probedata}{preprocessed probes x samples data matrix in absolute domain}

\item{verbose}{print intermediate messages}

\item{probe.parameters}{Optional. If probe.parameters are given,
         the summarization is based on these and model parameters are not
        estimated. A list. One element for each probeset with the following probe vectors:
        affinities, variances
Returns:}
}
\value{
List with two elements: abundance.table (summarized data matrix in absolute scale) and probe.parameters (RPA probe level parameter estimates)
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}

