% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{summarize.sum}
\alias{summarize.sum}
\title{Probeset summarization with SUM}
\usage{
summarize.sum(taxonomy, level, probedata, verbose = TRUE,
  downweight.ambiguous.probes = TRUE)
}
\arguments{
\item{taxonomy}{oligo - phylotype matching data.frame}

\item{level}{taxonomic level for the summarization.}

\item{probedata}{preprocessed probes x samples data matrix in absolute domain}

\item{verbose}{print intermediate messages}

\item{downweight.ambiguous.probes}{Downweight probes with multiple targets

Returns:}
}
\value{
List with two elements: abundance.table (summarized data matrix in absolute scale) and probe.parameters used in the calculations
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}

