% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{summarize_probedata}
\alias{summarize_probedata}
\title{Summarize phylogenetic microarray probe-level data from given input folder.}
\usage{
summarize_probedata(data.dir = NULL, probedata = NULL, taxonomy = NULL,
  level, method, probe.parameters = NULL)
}
\arguments{
\item{data.dir}{Data folder.}

\item{probedata}{probe-level data matrix}

\item{taxonomy}{probe taxonomy}

\item{level}{Summarization level}

\item{method}{Summarization method}

\item{probe.parameters}{Precalculater probe parameters. Optional.}
}
\value{
data matrix (taxa x samples)
}
\description{
Summarize phylogenetic microarray probe-level data from given input folder.
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}

