\name{tableprint}
\alias{tableprint}
\title{print a summary of data tables to screen.}
\usage{
  tableprint(y, typelabel = "table",
    dim1label = "rownames", dim2label = "colnames")
}
\arguments{
  \item{y}{the table to be printed.}

  \item{typelabel}{a character string indicating the class
  you want printed}

  \item{dim1label}{a character string indicating the rows
  dim name.}

  \item{dim2label}{a character string indicating the cols
  dim name.}
}
\description{
  This is internal. Not especially useful for others.
}
\keyword{internal}

