\name{tree.layout}
\alias{tree.layout}
\title{Returns a data frame defining segments to draw the phylogenetic tree.}
\usage{
  tree.layout(phylo, layout = "default", layout.ancestors =
  FALSE, ladderize = FALSE, align.seq.names = NA)
}
\description{
  This internal function is borrowed directly from the
  \code{ggphylo} package available on GitHub:
  \url{https://github.com/gjuggler/ggphylo}
}
\author{
  Gregory Jordan \email{gjuggler@gmail.com}
}
\seealso{
  This code is borrowed directly, with permission, from the
  not-yet-officially-released package, \code{ggphylo},
  currently only available from GitHub at:
  \url{https://github.com/gjuggler/ggphylo}
}
\keyword{internal}

