\name{UniFrac}
\alias{UniFrac}
\title{Alternative 'vectorized' implementation of the (unweighted) UniFrac calculation}
\usage{
  UniFrac(OTU, tree)
}
\arguments{
  \item{OTU}{otuTable in samples-by-species orientation}

  \item{tree}{object of class \code{phylo}}
}
\value{
  a distance matrix
}
\description{
  Alternative 'vectorized' implementation of the
  (unweighted) UniFrac calculation
}
\examples{
#
}
\seealso{
  wUniFrac unifrac
}
\keyword{UniFrac}

