% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{venn_phyloseq}
\alias{venn_phyloseq}
\title{Venn diagram of \code{\link{phyloseq-class}} object}
\usage{
venn_phyloseq(physeq, fact, min.nb.seq = 0, printValues = TRUE)
}
\arguments{
\item{physeq}{(Required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(Required): Name of the factor to cluster samples by modalities. 
Need to be in \code{physeq@sam_data}.}

\item{min.nb.seq}{(Default: 0)): minimum number of sequences by OTUs by samples 
to take into count this OTUs in this sample}

\item{printValues}{(logical) : Print (or not) the table of number of OTUs 
for each combination. If printValues is TRUE the object is not a ggplot object. 
Please use printValues = FALSE if you want to add ggplot function (cf example).}
}
\value{
A \code{\link{ggplot}}2 plot representing Venn diagramm of 
modalities of the argument \code{factor}
}
\description{
Venn diagram of \code{\link{phyloseq-class}} object
}
\examples{
data("enterotype")
\dontrun{
venn_phyloseq(enterotype, fact = 'SeqTech')
venn_phyloseq(enterotype, fact = 'ClinicalStatus')
venn_phyloseq(enterotype, fact = 'Nationality', printValues = F)
venn_phyloseq(enterotype, fact = 'ClinicalStatus', printValues = F) + scale_fill_hue()
venn_phyloseq(enterotype, fact = 'ClinicalStatus', printValues = F) + scale_fill_hue()
}

}
\author{
Adrien Taudiere
}
\seealso{
\code{\link[venneuler]{venneuler}}
}

