\docType{methods}
\name{otuTable<-}
\alias{assign-otuTable}
\alias{otuTable<-}
\alias{otuTable<-,otuSam,otuTable-method}
\alias{otuTable<-,otuSamTax,otuTable-method}
\alias{otuTable<-,otuSamTaxTree,otuTable-method}
\alias{otuTable<-,otuSamTree,otuTable-method}
\alias{otuTable<-,otuTable,otuTable-method}
\alias{otuTable<-,otuTree,otuTable-method}
\title{Assign to otuTable an object/value.}
\arguments{
  \item{x}{(Required). The object within which you will
  replace something}

  \item{value}{(Required). The value with which you will
  replace that thing in x.}
}
\description{
  Assign to otuTable an object/value.
}
\examples{
#
}

