\docType{methods}
\name{otuTable<-}
\alias{assign-otuTable}
\alias{otuTable<-}
\alias{otuTable<-,otuSam,otuTable-method}
\alias{otuTable<-,otuSamTax,otuTable-method}
\alias{otuTable<-,otuSamTaxTree,otuTable-method}
\alias{otuTable<-,otuSamTree,otuTable-method}
\alias{otuTable<-,otuTable,otuTable-method}
\alias{otuTable<-,otuTree,otuTable-method}
\title{Assign to otuTable an object/value.}
\description{
  Assign to otuTable an object/value.
}
\examples{
#
}

