\name{sample.names<-}
\alias{"assign-sample.names"}
\alias{"sample.names<-"}
\alias{sample.names<-}
\title{Assign sample names to an object.}
\description{
  This is typically used for subsetting an object to just
  those samples indicated by the character vector
  \code{value}.
}
\examples{
#
}
\seealso{
  prune.samples
}

