\name{species.names<-}
\alias{"assign-species.names"}
\alias{"species.names<-"}
\alias{species.names<-}
\title{Assign species names to an object.}
\description{
  This is typically used for subsetting an object to just
  those species indicated by the character vector
  \code{value}.
}
\examples{
#
}
\seealso{
  prune_species
}

