\docType{methods}
\name{tre<-}
\alias{assign-tre}
\alias{tre<-}
\alias{tre<-,phyloseq,phylo-method}
\title{Assign to tre an object/value.}
\usage{
  tre(x) <- value
}
\arguments{
  \item{x}{(Required). The object within which you will
  replace something}

  \item{value}{(Required). The value with which you will
  replace that thing in x.}
}
\description{
  Assign to tre an object/value.
}
\examples{
#
}

