\docType{data}
\name{data-ex1}
\alias{data-ex1}
\alias{ex1}
\title{(Data) Example data set 1 for the phyloseq package. (2011)}
\description{
  This is a small preliminary dataset of human gut
  microbial communities from subjects of different gender
  and diet. The exact meaning of diet and gender has been
  obscured because the larger study to which it belongs has
  not yet been published. For a more complete and
  already-published example dataset, try the
  \code{\link{enterotype}} data.
}
\examples{
## data(ex1)
}
\author{
  Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\references{
  \url{www.stanford.edu/~mcmurdie}
}
\keyword{data}
\keyword{datasets}

