\name{dist-class}
\alias{dist-class}
\title{An S4 placeholder for the basic \code{\link[stats]{dist}}ance matrix class.}
\description{
  See the \code{\link[ape]{ape}} package for details about
  this type of representation of a phylogenetic tree. It is
  used throught ape.
}
\seealso{
  \code{\link[stats]{dist}}

  \code{\link{setOldClass}}
}

