\name{[}
\alias{[}
\alias{[,otuTable-method}
\alias{[,phyloseq-method}
\alias{[,sampleMap-method}
\alias{[,taxonomyTable-method}
\title{Extract parts of otuTable}
\description{
  Extract parts of otuTable

  extract parts of sampleMap

  extract parts of taxonomyTable

  Generic extraction from higher-order object
}

