\name{[}
\alias{[}
\alias{[,otuTable,ANY,ANY,ANY-method}
\alias{[,phyloseq,ANY,ANY,ANY-method}
\alias{[,sampleMap,ANY,ANY,ANY-method}
\alias{[,taxonomyTable,ANY,ANY,ANY-method}
\title{Extract parts of otuTable}
\description{
  Extract parts of otuTable

  extract parts of sampleMap

  extract parts of taxonomyTable

  Generic extraction from higher-order object
}

