\name{[}
\alias{[}
\alias{[,otuTable-method}
\alias{[,phyloseq-method}
\alias{[,sampleData-method}
\alias{[,taxonomyTable-method}
\title{Extract parts of otuTable}
\description{
  Extract parts of otuTable

  extract parts of sampleData

  extract parts of taxonomyTable

  Generic extraction from higher-order object
}

