\name{[}
\alias{[}
\alias{[,otu_table,ANY-method}
\alias{[,phyloseq,ANY-method}
\alias{[,sample_data,ANY-method}
\alias{[,taxonomyTable,ANY-method}
\alias{[,XStringSet,character-method}
\title{Extract parts of otu_table}
\description{
  Extract parts of otu_table

  extract parts of sample_data

  extract parts of taxonomyTable

  Add name-character-based extraction method for XStringSet

  Generic extraction from phyloseq-class instance.
}

