\name{[}
\alias{[}
\alias{[,otu_table-method}
\alias{[,phyloseq-method}
\alias{[,sample_data-method}
\alias{[,taxonomyTable-method}
\title{Extract parts of otu_table}
\description{
  Extract parts of otu_table

  extract parts of sample_data

  extract parts of taxonomyTable

  Generic extraction from higher-order object
}

