\name{get.component.classes}
\alias{get.component.classes}
\title{Show the component objects classes and slot names.}
\value{
  a character vector of the component objects classes,
  where each element is named by the corresponding slot
  name in the higher-order phyloseq objects (objects
  containing more than 1 phyloseq data object).
}
\description{
  There are no arguments to this function. It returns a
  named character when called, which can then be used for
  tests of component data types, etc.
}
\examples{
#
#get.component.classes()
}

