% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/almostAllAccessors.R
\docType{methods}
\name{ntaxa}
\alias{ntaxa}
\alias{ntaxa,ANY-method}
\alias{ntaxa,XStringSet-method}
\alias{ntaxa,otu_table-method}
\alias{ntaxa,phylo-method}
\alias{ntaxa,phyloseq-method}
\alias{ntaxa,taxonomyTable-method}
\title{Get the number of taxa/species.}
\usage{
ntaxa(physeq)

\S4method{ntaxa}{ANY}(physeq)

\S4method{ntaxa}{phyloseq}(physeq)

\S4method{ntaxa}{otu_table}(physeq)

\S4method{ntaxa}{taxonomyTable}(physeq)

\S4method{ntaxa}{phylo}(physeq)

\S4method{ntaxa}{XStringSet}(physeq)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}, \code{\link{otu_table-class}},
 \code{\link{taxonomyTable-class}}, or
 \code{\link[ape]{phylo}}}
}
\value{
An integer indicating the number of taxa / species.
}
\description{
Get the number of taxa/species.
}
\examples{
data("esophagus")
ntaxa(esophagus)
phy_tree(esophagus)
ntaxa(phy_tree(esophagus))
}
\seealso{
taxa_names
}

