\name{pcoa}
\alias{pcoa}
\title{S3 class for ape-calculated MDS results}
\description{
  Nothing to import, because ape doesn't (yet) export this
  S3 class. We will define it here, but keep it internal.
  For the moment, its only use is for proper dispatch in
  our extensions to the scores S3 generic from vegan, for
  generic extraction of coordinates and possibly other
  features from any ordination results.
}
\keyword{internal}

