% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allClasses.R
\name{phylo-class}
\alias{phylo-class}
\title{An S4 placeholder of the main phylogenetic tree class from the ape package.}
\description{
See the \code{\link[ape]{ape}} package for details about this type of
representation of a phylogenetic tree. It is used throught ape.
}
\seealso{
\code{\link[ape]{phylo}}, \code{\link{setOldClass}}
}

