\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,otuSamTax,ANY-method}
\alias{plot,otuSamTree,ANY-method}
\alias{plot,otuTree,ANY-method}
\title{# #  Plot an otuTree object as a simple phylogenetic tree.
# #
# #}
\arguments{
  \item{x}{an \code{otuTree} object. # #}

  \item{...}{Additional plotting parameters, providede to
  \code{ape::plot.phylo()}}
}
\description{
  # # Plot an otuTree object as a simple phylogenetic tree.
  # # # #

  # # Plot a otuSam object as a taxaplot style barplot.

  # # Plot a otuSamTree object as an annotated phylogenetic
  tree.
}

