% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleData-class.R
\name{reconcile_categories}
\alias{reconcile_categories}
\title{Cleans absent levels in sample_data/data.frame.}
\usage{
reconcile_categories(DFSM)
}
\arguments{
\item{DFSM}{(Required). A \code{data.frame} or \code{sample_data} object that needs to be cleaned.}
}
\value{
A single \code{data.frame} object. Even if the input argument is a \code{sample_data},
 the return is a \code{data.frame}. Because this is intended to be used internally by
 the builder method, it cannot also call the builder function to re-build
 the cleaned \code{sample_data}.
}
\description{
This is used internally by the builder method, \code{\link{sample_data}}, to
ensure that the factors describing categorical variables in a data.frame or
sample_data object are free of extra levels that can plague downstream plots
analysis.
}
\examples{
# # # data(GlobalPatterns)
# # # SM <- sample_data(GlobalPatterns)
# # # DF <- data.frame(SM)
# # # DF <- data.frame(DF, col1=1:nrow(DF), col2=paste(1:nrow(DF), "t", sep=""))
# # # DF <- reconcile_categories(DF)
# # # SM <- sample_data(reconcile_categories(SM))
# # # sapply(DF, class)
# # # sapply(SM, class)
}
\keyword{internal}

