% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/almostAllAccessors.R
\docType{methods}
\name{sample_names}
\alias{sample_names}
\alias{sample_names,ANY-method}
\alias{sample_names,otu_table-method}
\alias{sample_names,phyloseq-method}
\alias{sample_names,sample_data-method}
\title{Get sample names.}
\usage{
sample_names(physeq)

\S4method{sample_names}{ANY}(physeq)

\S4method{sample_names}{phyloseq}(physeq)

\S4method{sample_names}{sample_data}(physeq)

\S4method{sample_names}{otu_table}(physeq)
}
\arguments{
\item{physeq}{(Required). A \code{\link{phyloseq-class}}, \code{\link{sample_data}},
 or \code{\link{otu_table-class}}.}
}
\value{
A character vector. The names of the samples in \code{physeq}.
}
\description{
Get sample names.
}
\examples{
data(esophagus)
sample_names(esophagus)
}
\seealso{
\code{\link{taxa_names}}, \code{\link{nsamples}}
}

