\docType{methods}
\name{show}
\alias{show}
\alias{show,otu_table-method}
\alias{show,phyloseq-method}
\alias{show,sample_data-method}
\alias{show,taxonomyTable-method}
\title{method extensions to show for phyloseq objects.}
\description{
  See the general documentation of
  \code{\link[methods]{show}} method for expected behavior.
}
\examples{
# data(GlobalPatterns)
# show(GlobalPatterns)
# GlobalPatterns
}
\seealso{
  \code{\link[methods]{show}}
}

