% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,otu_table-method}
\alias{show,otu_table-method}
\alias{show,phyloseq-method}
\alias{show,sample_data-method}
\alias{show,taxonomyTable-method}
\title{method extensions to show for phyloseq objects.}
\usage{
\S4method{show}{otu_table}(object)

\S4method{show}{sample_data}(object)

\S4method{show}{taxonomyTable}(object)

\S4method{show}{phyloseq}(object)
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\examples{
# data(GlobalPatterns)
# show(GlobalPatterns)
# GlobalPatterns
}
\seealso{
\code{\link[methods]{show}}
}

