\docType{methods}
\name{speciesAreRows}
\alias{speciesarerows}
\alias{speciesAreRows}
\alias{speciesAreRows,ANY-method}
\alias{speciesAreRows,otuTable-method}
\alias{speciesAreRows,phyloseq-method}
\title{Access speciesAreRows slot from otuTable objects.}
\arguments{
  \item{physeq}{(Required). \code{\link{phyloseq-class}},
  or \code{\link{otuTable-class}}.}
}
\value{
  A logical indicating the orientation of the otuTable.
}
\description{
  Access speciesAreRows slot from otuTable objects.
}
\seealso{
  \code{\link{otuTable}}
}

