\name{speciesarerows}
\alias{speciesarerows}
\alias{speciesAreRows}
\title{Access speciesAreRows slot from otuTable objects.}
\arguments{
  \item{object}{An object among the set of classes defined
  by the phyloseq package that are or contain an otuTable.}
}
\value{
  A logical indicating the orientation of the otuTable.
}
\description{
  Access speciesAreRows slot from otuTable objects.
}
\seealso{
  otuTable
}
\keyword{OTU}

