% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/almostAllAccessors.R
\docType{methods}
\name{taxa_are_rows}
\alias{taxa_are_rows}
\alias{taxa_are_rows,ANY-method}
\alias{taxa_are_rows,otu_table-method}
\alias{taxa_are_rows,phyloseq-method}
\title{Access taxa_are_rows slot from otu_table objects.}
\usage{
taxa_are_rows(physeq)

\S4method{taxa_are_rows}{ANY}(physeq)

\S4method{taxa_are_rows}{otu_table}(physeq)

\S4method{taxa_are_rows}{phyloseq}(physeq)
}
\arguments{
\item{physeq}{(Required). \code{\link{phyloseq-class}}, or \code{\link{otu_table-class}}.}
}
\value{
A logical indicating the orientation of the otu_table.
}
\description{
Access taxa_are_rows slot from otu_table objects.
}
\seealso{
\code{\link{otu_table}}
}

