\name{taxonomyTable-class}
\alias{taxonomyTable-class}
\title{An S4 class that holds taxonomic classification data as a character
matrix.}
\description{
  Row indices represent taxa, columns represent taxonomic
  classifiers.
}
\details{
  \describe{ \item{.Data}{This slot is inherited from the
  \code{\link{matrix}} class.} }
}

