\docType{methods}
\name{UniFrac}
\alias{"UniFrac,otuTable,phylo-method"}
\alias{"UniFrac,otuTree,ANY-method"}
\alias{UniFrac}
\title{Calculate unweighted UniFrac from an abundance matrix and a tree.}
\arguments{
  \item{OTU}{otuTable, or a more-complex object that
  contains an otuTable.}

  \item{tree}{object of class \code{phylo}, defined by ape
  package.}
}
\value{
  a distance matrix
}
\description{
  These methods depend on the picante and ape packages, as
  currently implemented.

  Calculate unweighted UniFrac from an abundance matrix and
  a tree.

  Calculate unweighted UniFrac from an abundance matrix and
  a tree.
}
\examples{
#
}
\seealso{
  wUniFrac unifrac
}

